backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ✅ ADD THIS - Gemini API (replaces OpenAI)
google-generativeai==0.3.2

# ✅ KEEP THIS - Free translation (no API key needed)
googletrans==4.0.0rc1

requests==2.32.3
python-dotenv==1.0.1
